+++ packages/pyroscope-flamegraph/src/FlameGraph/decode.ts
@@ -1,6 +1,37 @@
 import { Profile } from '@pyroscope/models/src';
 
-function deltaDiffWrapper(
+export function deltaDiffWrapperReverse(
+  format: Profile['metadata']['format'],
+  levels: Profile['flamebearer']['levels']
+) {
+  const mutableLevels = [...levels];
+
+  function deltaDiff(
+    lvls: Profile['flamebearer']['levels'],
+    start: number,
+    step: number
+  ) {
+    // eslint-disable-next-line no-restricted-syntax
+    for (const level of lvls) {
+      let total = 0;
+      for (let i = start; i < level.length; i += step) {
+        level[i] -= total;
+        total += level[i] + level[i + 1];
+      }
+    }
+  }
+
+  if (format === 'double') {
+    deltaDiff(mutableLevels, 0, 7);
+    deltaDiff(mutableLevels, 3, 7);
+  } else {
+    deltaDiff(mutableLevels, 0, 4);
+  }
+
+  return mutableLevels;
+}
+
+export function deltaDiffWrapper(
   format: Profile['metadata']['format'],
   levels: Profile['flamebearer']['levels']
 ) {
