+++ packages/pyroscope-flamegraph/src/convert/convertJaegerTraceToProfile.ts
@@ -2,43 +2,12 @@
 import groupBy from 'lodash.groupby';
 import map from 'lodash.map';
 import type { Profile, Trace, TraceSpan } from '@pyroscope/models/src';
+import { deltaDiffWrapperReverse } from '../FlameGraph/decode';
 
 interface Span extends TraceSpan {
   children: Span[];
   total: number;
   self: number;
-}
-
-// TODO: need to remove this ideally
-function deltaDiffWrapperReverse(
-  format: Profile['metadata']['format'],
-  levels: Profile['flamebearer']['levels']
-) {
-  const mutableLevels = [...levels];
-
-  function deltaDiff(
-    lvls: Profile['flamebearer']['levels'],
-    start: number,
-    step: number
-  ) {
-    // eslint-disable-next-line no-restricted-syntax
-    for (const level of lvls) {
-      let total = 0;
-      for (let i = start; i < level.length; i += step) {
-        level[i] -= total;
-        total += level[i] + level[i + 1];
-      }
-    }
-  }
-
-  if (format === 'double') {
-    deltaDiff(mutableLevels, 0, 7);
-    deltaDiff(mutableLevels, 3, 7);
-  } else {
-    deltaDiff(mutableLevels, 0, 4);
-  }
-
-  return mutableLevels;
 }
 
 export function convertJaegerTraceToProfile(trace: Trace): Profile {
@@ -110,7 +79,6 @@ export function convertJaegerTraceToProfile(trace: Trace): Profile {
 
   processNode(root, 0, 0);
 
-  // hack, need to remove this ideally
   resultFlamebearer.levels = deltaDiffWrapperReverse(
     'single',
     resultFlamebearer.levels
