+++ packages/pyroscope-flamegraph/src/index.tsx
@@ -4,6 +4,7 @@ import Flamegraph from './FlameGraph/FlameGraphComponent/Flamegraph';
 import { FlamegraphRenderer } from './FlamegraphRenderer';
 import { DefaultPalette } from './FlameGraph/FlameGraphComponent/colorPalette';
 import { convertJaegerTraceToProfile } from './convert/convertJaegerTraceToProfile';
+import { diffTwoProfiles } from './convert/diffTwoProfiles';
 
 export {
   Flamegraph,
@@ -11,4 +12,5 @@ export {
   FlamegraphRenderer,
   Box,
   convertJaegerTraceToProfile,
+  diffTwoProfiles,
 };
